Manifest builder added. Discoveryworker Interface added. QueueManager added. classes moved to folders. DiscoveryWorker changes. Program.cs changes AckPacket changes. CodeManager changes
